requirements.txt
awscli==1.18.132
boto3==1.14.55
botocore==1.17.55
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.13
django-allauth==0.42.0
django-redis==4.12.1
django-rest-auth==0.9.5
django-storages==1.10
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
drf-yasg==1.17.1
flex==6.14.1
gunicorn==20.0.4
httpie==2.2.0
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonpointer==2.0
MarkupSafe==1.1.1
mysqlclient==2.0.1
oauthlib==3.1.0
packaging==20.4
psycopg2-binary==2.8.5
pyasn1==0.4.8
Pygments==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
rsa==4.5
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
strict-rfc3339==0.7
uritemplate==3.0.1
urllib3==1.25.10
validate-email==1.3
whitenoise==5.1.0
